Avoid implicit bool conversions

+++ source/lac/trilinos_precondition_ml.cc
@@ -169,7 +169,7 @@ namespace TrilinosWrappers
                     TrilinosWrappers::global_index(domain_map, row) :
                     row;
                 distributed_constant_modes[d][row] =
-                  constant_modes[d][mode_index];
+                  static_cast<double>(constant_modes[d][mode_index]);
               }
           }
         (void)expected_mode_size;
